+++ api/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'org.springframework.boot' version '3.4.0'
+    id 'org.springframework.boot' version '3.1.2'
     id 'io.spring.dependency-management'
     id 'java'
 }
