+++ build.gradle
@@ -1,6 +1,6 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '3.4.0' apply false
+	id 'org.springframework.boot' version '3.1.2' apply false
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
@@ -32,7 +32,6 @@ subprojects {
 		annotationProcessor 'org.projectlombok:lombok'
 		testImplementation 'org.springframework.boot:spring-boot-starter-test'
 		testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-		implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0'
 
 	}
