+++ core/src/main/resources/application.yml
