+++ response/build.gradle
@@ -4,9 +4,9 @@ plugins {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation project(':core')
+    implementation project(':survey')
 
     runtimeOnly 'org.postgresql:postgresql'
 }
