+++ response/src/main/java/com/onboarding/response/config/ResponseConfig.java
@@ -1,0 +1,13 @@
+package com.onboarding.response.config;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EntityScan(basePackages = "com.onboarding.response.entity")
+@EnableJpaRepositories(basePackages = "com.onboarding.response.repository")
+public class ResponseConfig {
+
+}
