+++ response/src/main/java/com/onboarding/response/entity/ResponseDetail.java
@@ -1,0 +1,26 @@
+package com.onboarding.response.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class ResponseDetail {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "response_id")
+  private Response response;
+
+  private Long questionId;
+  private String questionType;
+  private boolean isRequired;
+  private String answer;
+}
