+++ response/src/main/java/com/onboarding/response/object/ResponseDetailObject.java
@@ -1,0 +1,10 @@
+package com.onboarding.response.object;
+
+public record ResponseDetailObject(
+    Long questionId,
+    String questionType,
+    boolean isRequired,
+    String answer
+) {
+
+}
