+++ response/src/main/java/com/onboarding/response/object/ResponseObject.java
@@ -1,0 +1,11 @@
+package com.onboarding.response.object;
+
+import java.util.List;
+
+public record ResponseObject(
+    Long surveyId,
+    String email,
+    List<ResponseDetailObject> responseDetails
+) {
+
+}
