+++ response/src/main/java/com/onboarding/response/repository/ResponseDetailRepository.java
@@ -1,0 +1,8 @@
+package com.onboarding.response.repository;
+
+import com.onboarding.response.entity.ResponseDetail;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ResponseDetailRepository extends JpaRepository<ResponseDetail, Long> {
+
+}
