+++ response/src/main/java/com/onboarding/response/repository/ResponseRepository.java
@@ -1,0 +1,10 @@
+package com.onboarding.response.repository;
+
+import com.onboarding.response.entity.Response;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface  ResponseRepository extends JpaRepository<Response, Long> {
+
+}
