+++ survey/src/main/java/com/onboarding/survey/enums/Operation.java
@@ -1,0 +1,5 @@
+package com.onboarding.survey.enums;
+
+public enum Operation {
+  ADD, UPDATE, DELETE
+}
