+++ survey/src/main/java/com/onboarding/survey/enums/QuestionType.java
@@ -1,0 +1,8 @@
+package com.onboarding.survey.enums;
+
+public enum QuestionType {
+  SHORT_ANSWER,
+  LONG_ANSWER,
+  SINGLE_CHOICE,
+  MULTIPLE_CHOICE
+}
