+++ survey/src/main/java/com/onboarding/survey/repository/QuestionRepository.java
@@ -1,0 +1,10 @@
+package com.onboarding.survey.repository;
+
+import com.onboarding.survey.entity.Question;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuestionRepository extends JpaRepository<Question, Long> {
+
+}
