+++ survey/src/main/java/com/onboarding/survey/repository/SurveyRepository.java
@@ -1,0 +1,10 @@
+package com.onboarding.survey.repository;
+
+import com.onboarding.survey.entity.Survey;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
+
+}
