--- survey/src/main/java/com/onboarding/survey/survey/enums/Operation.java
@@ -1,5 +1,0 @@
-package com.onboarding.survey.survey.enums;
-
-public enum Operation {
-  ADD, UPDATE, DELETE
-}
