--- survey/src/main/java/com/onboarding/survey/survey/enums/QuestionType.java
@@ -1,8 +1,0 @@
-package com.onboarding.survey.survey.enums;
-
-public enum QuestionType {
-  SHORT_ANSWER,
-  LONG_ANSWER,
-  SINGLE_CHOICE,
-  MULTIPLE_CHOICE
-}
