--- survey/src/main/java/com/onboarding/survey/survey/repository/QuestionRepository.java
@@ -1,10 +1,0 @@
-package com.onboarding.survey.survey.repository;
-
-import com.onboarding.survey.survey.entity.Question;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-
-}
