--- survey/src/main/java/com/onboarding/survey/survey/service/read/ReadQuestionService.java
@@ -1,19 +1,0 @@
-package com.onboarding.survey.survey.service.read;
-
-import com.onboarding.core.global.exception.CustomException;
-import com.onboarding.core.global.exception.enums.ErrorCode;
-import com.onboarding.survey.survey.entity.Question;
-import com.onboarding.survey.survey.repository.QuestionRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ReadQuestionService {
-
-  private final QuestionRepository questionRepository;
-
-  public Question getQuestionById(Long questionId) {
-    return questionRepository.findById(questionId).orElseThrow(() -> new CustomException(ErrorCode.ENTITY_NOT_FOUND));
-  }
-}
