+++ survey/src/test/java/com/onboarding/survey/survey/facade/SurveyFacadeTest.java
@@ -6,12 +6,13 @@ import static org.mockito.Mockito.when;
 
 import com.onboarding.core.global.exception.CustomException;
 import com.onboarding.core.global.exception.enums.ErrorCode;
-import com.onboarding.survey.survey.entity.Question;
-import com.onboarding.survey.survey.entity.Survey;
-import com.onboarding.survey.survey.enums.QuestionType;
-import com.onboarding.survey.survey.service.create.CreateSurveyService;
-import com.onboarding.survey.survey.service.read.ReadQuestionService;
-import com.onboarding.survey.survey.service.read.ReadSurveyService;
+import com.onboarding.survey.entity.Question;
+import com.onboarding.survey.entity.Survey;
+import com.onboarding.survey.enums.QuestionType;
+import com.onboarding.survey.facade.SurveyFacade;
+import com.onboarding.survey.service.create.CreateSurveyService;
+import com.onboarding.survey.service.read.ReadQuestionService;
+import com.onboarding.survey.service.read.ReadSurveyService;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
